+++ src/components/NewRestaurantModal/NewRestaurantModal.ts
@@ -5,10 +5,11 @@ import { IRestaurant, DistanceOrPlaceholder, Category, DistanceNumeric } from '@
 import './NewRestaurantModal.css';
 import AllRestaurantApp from '../AllRestaurantApp';
 import { dom } from '@/util/dom';
+import NewRestaurantForm from './NewRestaurantForm';
 
 class NewRestaurantModal extends BasicModal {
   #title: HTMLHeadingElement;
-  #form: HTMLFormElement;
+  #form: NewRestaurantForm;
 
   constructor() {
     super();
@@ -27,26 +28,16 @@ class NewRestaurantModal extends BasicModal {
   }
 
   closeModal() {
-    this.#invisibleErrorMessage();
+    this.#form.invisibleErrorMessage();
     this.classList.remove('modal--open');
-  }
-
-  getForm() {
-    return this.#form;
-  }
-
-  #invisibleErrorMessage() {
-    dom.getElement<HTMLElement>(this, '.category-select > .error').classList.add('invisible');
-    dom.getElement<HTMLElement>(this, '.distance-select > .error').classList.add('invisible');
-    dom.getElement<HTMLElement>(this, '.name-input-box > .error').classList.add('invisible');
   }
 
   #setSubmitEvent() {
     this.#form.addEventListener('submit', (e) => {
       e.preventDefault();
-      this.#invisibleErrorMessage();
-      const { name, distance, category, description, link } = this.#getValues();
-      if (this.#validateRequiredValues(category, distance, name)) return;
+      this.#form.invisibleErrorMessage();
+      const { name, distance, category, description, link } = this.#form.getValues();
+      if (this.#form.validateRequiredValues(category, distance, name)) return;
 
       const distanceNumeric = distance as DistanceNumeric;
       const categoryOnly = category as Category;
@@ -57,51 +48,15 @@ class NewRestaurantModal extends BasicModal {
         ...(description && { description }),
         ...(link && { link }),
       };
-      const DBService = new RestaurantDBService();
-      DBService.add(newRestaurant);
+      new RestaurantDBService().add(newRestaurant);
 
       this.#rerenderApp();
       this.closeModal();
     });
   }
 
-  #getValues(): {
-    name: string;
-    distance: number;
-    category: string;
-    description: string;
-    link: string;
-  } {
-    const name: string = (this.#form.elements.namedItem('name') as HTMLInputElement).value;
-    const distance = Number(
-      (this.#form.elements.namedItem('distance') as HTMLSelectElement)
-        .value as DistanceOrPlaceholder,
-    );
-
-    const category = (this.#form.elements.namedItem('category') as HTMLSelectElement).value;
-    const description = (this.#form.elements.namedItem('description') as HTMLInputElement).value;
-    const link = (this.#form.elements.namedItem('link') as HTMLInputElement).value;
-    return { name, distance, category, description, link };
-  }
-
-  #validateRequiredValues(category: string, distance: number, name: string | null) {
-    const isNotValidCategory = category === '선택해주세요';
-    const isNotValidDistance = Number.isNaN(distance);
-    const isNotValidName = !name;
-    if (isNotValidCategory) {
-      dom.getElement(this, '.category-select > .error').classList.remove('invisible');
-    }
-    if (isNotValidDistance) {
-      dom.getElement(this, '.distance-select > .error').classList.remove('invisible');
-    }
-    if (isNotValidName) {
-      dom.getElement(this, '.name-input-box > .error').classList.remove('invisible');
-    }
-    return isNotValidCategory || isNotValidDistance || isNotValidName;
-  }
-
   #rerenderApp() {
-    (this.parentElement as AllRestaurantApp).render();
+    dom.getElement<AllRestaurantApp>(document.body, '[is="all-restaurant-app"]').render();
   }
 }
 export default NewRestaurantModal;
